Optimize rendering
Depth is now handled by a depth buffer that is integrated into the main render target, instead of a custom render target.
Shadows are now handled in a separate render pass using the stencil buffer to mark which pixels have already been shaded.

This eliminates the terrain depth render pass completely and allows the terrain color render pass to be batched.
Overall performance improvement is from a few dozen % to up to 300% depending on what is on the screen and what the user is doing.